+++ packages/credential-format/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gramota/credential-format
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @gramota/core@0.2.1
+
 ## 0.1.2
 
 ### Patch Changes