+++ packages/dcql/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gramota/dcql
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @gramota/core@0.2.1
+  - @gramota/sd-jwt@0.2.2
+
 ## 0.1.2
 
 ### Patch Changes