+++ packages/holder/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @gramota/holder
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @gramota/jose@0.3.0
+  - @gramota/core@0.2.1
+  - @gramota/oid4vci@0.3.1
+  - @gramota/oid4vp@0.2.2
+  - @gramota/sd-jwt@0.2.2
+  - @gramota/presentation-exchange@0.1.3
+
 ## 0.1.2
 
 ### Patch Changes