+++ packages/status-list/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @gramota/status-list
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @gramota/jose@0.3.0
+  - @gramota/core@0.2.1
+  - @gramota/sd-jwt@0.2.2
+
 ## 0.1.2
 
 ### Patch Changes