+++ packages/trust/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gramota/trust
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @gramota/jose@0.3.0
+  - @gramota/core@0.2.1
+
 ## 0.1.2
 
 ### Patch Changes